Set size, stride, and offset of functional tensor.
Fix: #120585

This PR modifies at::functionalization::impl::set_sizes_strides_offset, calling
TensorImpl::set_sizes_and_strides not only for the FunctionalTensorWrapper tensor, but
also for its inner value (which is also a tensor).

The rationale is that FunctionalTensorWrapper actually overrides its metadata getter
methods, forwarding them to its inner value. Therefore, whenever we try to set that
information for a FunctionalTensorWrapper, it makes sense to actually set it for its
inner value.

This also updates the PyTorch/XLA pin to https://github.com/pytorch/xla/pull/8032 so that
the XLA tests pass.

ghstack-source-id: 4fadaeaf9fd89b67ce310b6388af500a0e4f3431
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135237