[POC] fast eager-mode HOPs
Based off of Animesh's invoke_subgraph POC.

This makes it so that we do not materialize a forward or backward graph
at DispatchKey::Autograd, instead we construct a wrapper function and
then do a re-dispatch. If there is a ProxyTorchDispatchMode later on,
then we will materialize the graph at that point.

Materialization of the forward graph *does not* require us to
materialize the backward graph (or make decisions on the backward
tangents (gradients)!)

ghstack-source-id: e1f2864524f856682612be20b6beb363c133c391
Pull Request resolved: https://github.com/pytorch/pytorch/pull/137360