[Pipelining] add schedule simulator and chrometrace dump
Schedule simulator is useful for detecting hangs in schedules and
validating that they won't hang.  It also inserts bubbles (None actions)
at any timestep where a rank can not enqueue its next action due to
unmet dependencies, which can serve as a rough metric for schedule
efficiency.  The output can be visualized.  The simulator expects a full
comm + compute schedule as input.

Chrometrace dump is a basic visualization utility.  It currently just
renders one 'process' per rank, and lets users visualize the schedule in
a UI instead of as text.

ghstack-source-id: e5aede7ac1946365a2bacae0ebaa7c513c805ee0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/138134