[logging] Superficial use of new metrics_context in CompilationMetrics logging
Summary: This change makes use of the new metrics accounting utility to record CompilationMetrics. This is a first step and not a deep integration: whereever we were previously creating a CompilationMetrics object to log, we instead update the current metrics_context and rely on that contextmanager to call record_compilation_metrics when it exits. In a subsequent change, we'll add usage of the 'timed' contextmanager to collect various compile times.

Test Plan: unit test

ghstack-source-id: 3a1a5f33e7da2eeb6229a36ef1f3e6207a1b082f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139260