Add repro instructions to fx_graph_runnable.py
This PR adds some instructions for how to add a TARGETS file to run the
fx_graph_runnable script. I'm planning to add some followups that will
add additional imports for custom ops and use autodeps to get the
dependencies, but I figure this PR is an easy first step.

Test Plan:
- pytest test/dynamo/test_structured_trace.py
- Does anyone have suggestions for how to test this?

ghstack-source-id: 5f3ed2ead69ad4f02cb462c9388ec0a9b16cbc99
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139481