[rfc][dynamo] "skip_guard_eval_unsafe" API for power users
We have spent quite some time this year on improving guard performance
and soundness. Nevertheless, guards STILL take time. We have seen
multiple requests/evidences from power users where they want to have
almost 0% guard overhead. First, we saw this in vLLM where even 1%
ovehread is bad. And recently we saw this in hqq (low precision LLM
generation) - https://github.com/pytorch/pytorch/issues/138386. To put
some numbers for perspecitve, low precision LLM infernece reaches around
250 tokens/second, i.e, each token takes a mere 4 milliseconds. If guard
overhead is even 200 us, its still 5% overhead in total.

Here, users ask - "we can guarantee that there will no more
recompilations in the steady state, give us the lowest guard overhead"

A must-have consideration is to support fast inference where the model
has recompiled (could be because of dynamism, or just tensor dtype
change in the case of hqq). So, we still have to run the guards to
figure out which compiled graph to run.

What we need is the "minimal set of differentiating guards" - i.e.
minimals set of guards that we can run to choose the compiled graph.
Note that this works ONLY with the assumption that users really
guarantee no more recompilation scenarios (no more mutations, no more
dynamism after the model has been warmed up).

When we designed C++ guards, Ed and Voz suggested to use Trie-structure
to directly represent this "diff guard set". But due to complexity, we
went for tree structure and relied on a GuardManager state -
"fail_count" - to fail fast. I realized that we can rely on this
"fail_count" to find the diff guard set.

If we recompile, this means that all the cache line guard eval check_fns
have failed. Whenever a guard check_fn fails, we increment the counter
in the tree node to increase the weightage of that node. If we just want
to run the "guard diff set", we just have to run only those nodes in the
tree which have fail_count > 0.

This PR relies on this observation to introduce a new API -
"skip_guard_eval". The idea is that user will warm up their model with
torch.compile, and the run the steady state with this API. This
API goes through the existing cache lines for the intercepted code
object but only runs the diff guard set. This dramatically reduces the
guard overhead. In case, all guards fail, we fall back to eager (however
if this happens then user is violating the assumption, so we should
perhaps hard error).

I tried the hqq model in
https://github.com/pytorch/pytorch/issues/138386. With very small
changes in the user code (share PR), I see the throughput increase from 160
tokens/sec to 174 tokens/sec.

ghstack-source-id: 9332fde8700fee7216a99e582e4b5fbf494aa81c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139038