[pipelining] RNG seed management for PP init
Add new 'seeded_module_init' API to PipelineScheduleMulti
- note: skipped for PipelineScheduleSingle for now. Could add it there
  too, but i want to delete that class and unify

Example:

```
schedule.seeded_module_init(
   # required arg, can be any fn that accepts 'module' as arg;
   # can be ModuleClass.init_weights
   my_initializer_fn,

   # optional args
   mode="serial",
   seed=1234
)
```

Implement both 'serial' and 'parallel' options described in #139304

'parallel' mode sets a unique seed per rank, but does not worry about
parity with non-PP initialization.  It is fastest becuase it
parallelizes initialization and uses no communication.

'serial' mode does initialization one stage at a time, communicating the
RNG states from rank to rank as needed so that the resulting
initialization is equivalent to non-PP initialization where the whole
model is initialized on one rank.

The overall initialization is delegated to the 'schedule' since it knows which
stages it owns, which ranks they are on, and what order they go in.
But the per-stage initialization is handled by a user-passed function
that can do anything it wants to the given module owned by that stage.

ghstack-source-id: 35ce7ed77fe9d01e39da98eb82a558c961637a88
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139445