[dynamo] Add a `DynamoFrameType` type above Python frame object
This patch introduces a `DynamoFrameType` to serve as a layer between
Dynamo and different versions of Python frame object. In
`DynamoFrameType`, we only register attributes Dynamo cares about (e.g.,
`f_code`, `f_locals`, etc.

This will be helpful when it comes to adding new attributes to this
`DynamoFrameType`, or dealing with Python version changes.

ghstack-source-id: 36f11e5f0abf39105389c0e2c39de89bff0c4e56
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140330