[C10D] call destroy_process_group after MultiProcess tests
Faced with an annoying string of warnings like this when running tests,

My choices seem to be (1) call destroy_process_group() at the end of
each test fn, (2) do this in some wrapper, (3) do it in the base test
class.

Since tests in MultiProcessTestCase are responsible for calling
init_process_group themselves, they should also be responsible for
calling destroy (or at least method (3) would be asymmetric and may
result in double-destroy).

But it doesn't feel worth it to go add a destroy call manually to each
test, and try/except for a possible second destroy call seems like a
happy middle ground.

Note: tests that want to ensure that destroy runs cleanly can and should
still call destroy _inside_ the test, and this change does not affect
that.

ghstack-source-id: 6a4168b5308bc8c06759cfae710a1755dd97033c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140820