Don't throw on setting metric twice if the metric value hasn't changed
This is causing FbFxGraphRemoteCache.__init__ to no longer be idempotent, i.e. only safe to call once per compile. AOTAutogradCache initializes a new remote cache for the forward and the backward.

Technically, we could make AOTAutogradCache smart and globally thread through a single FbFxGraphRemoteCache everywhere. But there's no reason to do so, as this class is just the handle to access the cache. Plus, it's very brittle for FbFxGraphRemoteCache to not be safe to call multiple times.

Differential Revision: [D66502138](https://our.internmc.facebook.com/intern/diff/D66502138/)

ghstack-source-id: 255502771
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141582