[dynamo] Fix side-effect handling for pre-existing `collections.deque`
Previously we never replayed side effects to `DequeVariable` with a
source; the bug was already in the `test_deque_input` test, but went
unnoticed because we didn't check the deque objects.

This patch adds limited but practical support for this (see comments in
`side_effects.py` for why limited), and updates the deque tests to check
for this.

ghstack-source-id: 2b49768dab9e951853817c6d43e461f80b240bf0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141714